SAW-2401 Update kpi design and add execute mocks

+++ saw-web/src/main/javascript/app/modules/observe/components/add-widget/widget-metric/widget-metric.component.ts
@@ -28,7 +28,7 @@ export class WidgetMetricComponent implements OnInit {
           {
             columnName: 'AVAILABLE_MB',
             displayName: 'Available MB',
-            type: 'double'
+            type: 'integer'
           }
         ];
